WIP. Migrating bot app to ESM
