+++ app/scripts/create-user.js
@@ -10,7 +10,7 @@ import inquirer from 'inquirer'
 
 import redis from '../services/redis.js'
 import { REDIS } from '../services/constants.js'
-import { dirname } from '../utils/dirname.js'
+import { dirname } from '../services/utils.js'
 
 const figletPromise = bluebird.promisify(figlet)
